Ra 574 bug fix (#364)
* Attempt to fix bug

* If nothing to scan, then just exit

* Added TODO
