+++ pipeline-runner/steps/build_acr.go
@@ -7,32 +7,21 @@ import (
 	"time"
 
 	"github.com/equinor/radix-operator/pipeline-runner/model"
-	"github.com/equinor/radix-operator/pkg/apis/pipeline"
 	"github.com/equinor/radix-operator/pkg/apis/utils"
 	"github.com/equinor/radix-operator/pkg/apis/utils/git"
 
 	"github.com/equinor/radix-operator/pkg/apis/kube"
 	v1 "github.com/equinor/radix-operator/pkg/apis/radix/v1"
-	log "github.com/sirupsen/logrus"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-const (
-	azureServicePrincipleSecretName = "radix-sp-acr-azure"
-	multiComponentImageName         = "multi-component"
-)
+const azureServicePrincipleSecretName = "radix-sp-acr-azure"
 
 type void struct{}
 
 var member void
-
-type componentType struct {
-	name           string
-	context        string
-	dockerFileName string
-}
 
 func createACRBuildJob(rr *v1.RadixRegistration, ra *v1.RadixApplication, containerRegistry string, pipelineInfo *model.PipelineInfo, buildSecrets []corev1.EnvVar) (*batchv1.Job, error) {
 	appName := rr.Name
@@ -121,8 +110,7 @@ func createACRBuildContainers(containerRegistry, appName string, pipelineInfo *m
 	}
 
 	distinctBuildContainers := make(map[string]void)
-	componentImages := getComponentImages(appName, containerRegistry, imageTag, components)
-	for _, componentImage := range componentImages {
+	for _, componentImage := range pipelineInfo.ComponentImages {
 		if !componentImage.Build {
 			// Nothing to build
 			continue
@@ -199,77 +187,5 @@ func createACRBuildContainers(containerRegistry, appName string, pipelineInfo *m
 		containers = append(containers, container)
 	}
 
-	pipelineInfo.ComponentImages = componentImages
 	return containers
-}
-
-func getComponentImages(appName, containerRegistry, imageTag string, components []v1.RadixComponent) map[string]pipeline.ComponentImage {
-	// First check if there are multiple components pointing to the same build context
-	buildContextComponents := make(map[string][]componentType)
-
-	for _, c := range components {
-		if c.Image != "" {
-			// Using public image. Nothing to build
-			continue
-		}
-
-		componentSource := getDockerfile(c.SourceFolder, c.DockerfileName)
-		components := buildContextComponents[componentSource]
-		if components == nil {
-			components = make([]componentType, 0)
-		}
-
-		components = append(components, componentType{c.Name, getContext(c.SourceFolder), getDockerfileName(c.DockerfileName)})
-		buildContextComponents[componentSource] = components
-	}
-
-	componentImages := make(map[string]pipeline.ComponentImage)
-
-	// Gather pre-built or public images
-	for _, c := range components {
-		if c.Image != "" {
-			componentImages[c.Name] = pipeline.ComponentImage{Build: false, Scan: false, ImageName: c.Image, ImagePath: c.Image}
-		}
-	}
-
-	// Gather build containers
-	numMultiComponentContainers := 0
-	for _, components := range buildContextComponents {
-		var imageName string
-
-		if len(components) > 1 {
-			log.Infof("Multiple components points to the same build context")
-			imageName = multiComponentImageName
-
-			if numMultiComponentContainers > 0 {
-				// Start indexing them
-				imageName = fmt.Sprintf("%s-%d", imageName, numMultiComponentContainers)
-			}
-
-			numMultiComponentContainers++
-		} else {
-			imageName = components[0].name
-		}
-
-		buildContainerName := fmt.Sprintf("build-%s", imageName)
-
-		// A multi-component share context and dockerfile
-		context := components[0].context
-		dockerFile := components[0].dockerFileName
-
-		// Set image back to component(s)
-		for _, c := range components {
-			componentImages[c.name] = pipeline.ComponentImage{
-				ContainerName: buildContainerName,
-				Context:       context,
-				Dockerfile:    dockerFile,
-				ImageName:     imageName,
-				ImagePath:     utils.GetImagePath(containerRegistry, appName, imageName, imageTag),
-				Build:         true,
-				Scan:          true,
-			}
-		}
-	}
-
-	return componentImages
 }
